researcher/pyproject.toml
[tool.poetry]
name = "researcher"
version = "0.1.0"
description = "Ancient art researcher"
authors = ["Vikram Pawar <pvikram035@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = { extras = ["tools"], version = "^0.41.1" }
google-auth = "^2.32.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.139.0"
flask = "^3.0.3"
google-cloud-firestore = "^2.17.0"
google-cloud-storage = "^2.18.2"
firecrawl-py = "^0.0.20"
langchain-groq = "^0.1.9"
fastembed = "^0.3.4"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
redis = "researcher.manage_redis:start_redis_server"
researcher = "researcher.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
